Rename referenced object test file to match other test files
