+++ sbol/test/test_referencedobject.py
